+++ packages/instrumentation-anthropic/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.17.0](https://github.com/traceloop/openllmetry-js/compare/v0.16.2...v0.17.0) (2025-08-18)
+
+**Note:** Version bump only for package @traceloop/instrumentation-anthropic
+
 # [0.16.0](https://github.com/traceloop/openllmetry-js/compare/v0.15.0...v0.16.0) (2025-08-14)
 
 **Note:** Version bump only for package @traceloop/instrumentation-anthropic