+++ packages/instrumentation-langchain/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.17.0](https://github.com/traceloop/openllmetry-js/compare/v0.16.2...v0.17.0) (2025-08-18)
+
+### Features
+
+- **langchain:** implement callback-based instrumentation with auto-injection ([#649](https://github.com/traceloop/openllmetry-js/issues/649)) ([fe78b1b](https://github.com/traceloop/openllmetry-js/commit/fe78b1b512d04f8e33e2f1886465e08870a1b56f))
+
 # [0.16.0](https://github.com/traceloop/openllmetry-js/compare/v0.15.0...v0.16.0) (2025-08-14)
 
 **Note:** Version bump only for package @traceloop/instrumentation-langchain